Tradeverse-AI/requirements.txt
fastapi
uvicorn
python-dotenv
pinecone
sentence-transformers
requests
httpx
yfinance
pandas
numpy
transformers
torch
